move theme from content to themedesigner
