+++ packages/react-components/theme-designer/src/components/Content/Content.tsx
@@ -6,10 +6,13 @@ import { Demo } from '../Demo/Demo';
 import { Palette } from '../Palette/Palette';
 import { TokenBoxes } from '../TokenBoxes/TokenBoxes';
 
-import { createLightTheme, createDarkTheme, BrandVariants } from '@fluentui/react-theme';
+import { Theme, BrandVariants } from '@fluentui/react-theme';
 
 export interface ContentProps {
   className?: string;
+  brand: BrandVariants;
+  darkTheme: Theme;
+  lightTheme: Theme;
 }
 
 const useStyles = makeStyles({
@@ -23,41 +26,18 @@ const useStyles = makeStyles({
   },
 });
 
-// this data is temporary and will eventually be pulled from current theme
-const brand: BrandVariants = {
-  10: `#061724`,
-  20: `#082338`,
-  30: `#0a2e4a`,
-  40: `#0c3b5e`,
-  50: `#0e4775`,
-  60: `#0f548c`,
-  70: `#115ea3`,
-  80: `#0f6cbd`,
-  90: `#2886de`,
-  100: `#479ef5`,
-  110: `#62abf5`,
-  120: `#77b7f7`,
-  130: `#96c6fa`,
-  140: `#b4d6fa`,
-  150: `#cfe4fa`,
-  160: `#ebf3fc`,
-};
-
 export const Content: React.FC<ContentProps> = props => {
   const styles = useStyles();
   const [isDark, setIsDark] = React.useState<boolean>(false);
 
   const toggleTheme = React.useCallback(() => setIsDark(!isDark), [isDark, setIsDark]);
 
-  const LightTheme = createLightTheme(brand);
-  const DarkTheme = createDarkTheme(brand);
-
-  const theme = isDark ? DarkTheme : LightTheme;
+  const theme = isDark ? props.darkTheme : props.lightTheme;
 
   return (
     <FluentProvider theme={theme}>
       <div className={mergeClasses(styles.root, props.className)}>
-        <Palette brandColors={brand} />
+        <Palette brandColors={props.brand} />
         <Demo theme={theme} />
         <Divider />
         <TokenBoxes theme={theme} isDark={isDark} toggleTheme={toggleTheme} />
